feat: split embed handling to different files
